+++ instrumentation/github.com/Shopify/sarama/otelsarama/producer.go
@@ -26,7 +26,7 @@ import (
 	"go.opentelemetry.io/otel/codes"
 
 	"go.opentelemetry.io/otel/attribute"
-	semconv "go.opentelemetry.io/otel/semconv/v1.12.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
 	"go.opentelemetry.io/otel/trace"
 )
 
@@ -287,7 +287,7 @@ func startProducerSpan(cfg config, version sarama.KafkaVersion, msg *sarama.Prod
 	attrs := []attribute.KeyValue{
 		semconv.MessagingSystemKey.String("kafka"),
 		semconv.MessagingDestinationKindTopic,
-		semconv.MessagingDestinationKey.String(msg.Topic),
+		semconv.MessagingDestinationNameKey.String(msg.Topic),
 		semconv.MessagingMessagePayloadSizeBytesKey.Int(msgPayloadSize(msg, version)),
 		semconv.MessagingOperationKey.String("send"),
 	}
@@ -308,7 +308,7 @@ func startProducerSpan(cfg config, version sarama.KafkaVersion, msg *sarama.Prod
 func finishProducerSpan(span trace.Span, partition int32, offset int64, err error) {
 	span.SetAttributes(
 		semconv.MessagingMessageIDKey.String(strconv.FormatInt(offset, 10)),
-		semconv.MessagingKafkaPartitionKey.Int64(int64(partition)),
+		semconv.MessagingKafkaDestinationPartitionKey.Int64(int64(partition)),
 	)
 	if err != nil {
 		span.SetStatus(codes.Error, err.Error())